fix: rename utils to index file
