+++ apps/web/cypress/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "types": ["cypress"]
   },
-  "include": ["../../../node_modules/cypress", "./**/*.ts", "utils/utils.ts", "plugins/index.js"]
+  "include": ["../../../node_modules/cypress", "./**/*.ts"]
 }
