+++ apps/web/cypress/utils/index.ts
